# sdk/log: Fix BenchmarkLoggerNewRecord to not drop attributes